Fix statistics graph paint bugs (UI hang + first-paint NPE)
From the custom-paint audit:
- GraphHistory: the vertical-grid loop used a byte counter that
  overflowed past 127 (120 + 20 -> -116), so it never terminated once
  the window was >= 128px wide, hanging the UI in the paint handler when
  the graph-history window opened. Use an int counter (the sibling
  horizontal loop already did).
- GraphCanvas: paintControl assumed controlResized (the only setter of
  needNewBuffer) always ran before the first paint; modern SWT doesn't
  guarantee that, leaving imageBuffer null and new GC(null) throwing.
  Also build the buffer when it is null.

Deferred to ToDo: chunk-bar blur on >100% HiDPI displays (ImageData
round-trip loses device resolution) — needs a scaled display to verify.

+++ sancho-src/src/sancho/view/statistics/GraphHistory.java
@@ -81,7 +81,10 @@ public class GraphHistory implements PaintListener {
          var5.fillRectangle(0, 0, this.shell.getClientArea().width, this.shell.getClientArea().height);
          var5.setForeground(this.gridColor);
 
-         for (byte var6 = 0; var6 < var3; var6 += 20) {
+         for (int var6 = 0; var6 < var3; var6 += 20) {
+            // var6 must be int: as a byte it overflowed past 127 (120 + 20 -> -116)
+            // and the loop never terminated once the width was >= 128px, hanging the
+            // UI when the graph-history window opened.
             var5.drawLine(var6, 0, var6, var2 + 1);
          }
